fix output annotation

+++ src/funpay_undercutter/models.py
@@ -84,7 +84,7 @@ class Offer(Base):
         return f'{self.seller_name: <18} {round(float(self.price) / 1000, 3): <5}₽ {self.amount: <10} ' \
                f'{self.server.name: <20} {self.side_name: <8} {self.game.name: <8}'
 
-    def price_for_save(self, commission: float) -> int:
+    def price_for_save(self, commission: float) -> float:
         return (self.price - self.price * commission / 100)/1000
